+++ test/Interop/Cxx/class/constructors-executable.swift
@@ -1,4 +1,4 @@
-// RUN: %target-run-simple-swift(-I %S/Inputs/ -Xfrontend -enable-experimental-cxx-interop)
+// RUN: %target-run-simple-swift(-I %S/Inputs/ -Xfrontend -enable-experimental-cxx-interop -Xfrontend -disable-availability-checking)
 //
 // REQUIRES: executable_test
 
@@ -68,6 +68,24 @@ CxxConstructorTestSuite.test("MoveConstructorWithOneParamWithDefaultArg") {
   let instance2 = instance1
   let instance3 = MoveConstructorWithOneParameterWithDefaultArg(5)
   expectTrue(instance2.value + instance3.value >= 10)
+}
+
+CxxConstructorTestSuite.test("ImportStaticFactoryAsInitializer") {
+  let x = ImportWithCtor()
+  expectEqual(x.param1, 0)
+  expectEqual(x.param2, 0)
+  let y = x
+  let z = ImportWithCtor(1)
+  expectEqual(z.param1, 1)
+  expectEqual(z.param2, 0)
+  let z2 = ImportWithCtor(2, 3)
+  expectEqual(z2.param1, 2)
+  expectEqual(z2.param2, 3)
+  let z3 = ImportWithCtor(2, 3, 4)
+  expectEqual(z3.param1, 2)
+  expectEqual(z3.param2, 3)
+  let v = Value(x: 2)
+  expectEqual(v.getX(), 2)
 }
 
 runAllTests()
